Rename modules into a more proper name
