+++ .idea/gradle.xml
@@ -19,7 +19,7 @@
             <option value="$PROJECT_DIR$/data/network" />
             <option value="$PROJECT_DIR$/domain" />
             <option value="$PROJECT_DIR$/feature" />
-            <option value="$PROJECT_DIR$/feature/home" />
+            <option value="$PROJECT_DIR$/feature/news" />
           </set>
         </option>
       </GradleProjectSettings>
