+++ app/src/main/java/com/mnhyim/warta/utils/HomeNavigationBarDestinations.kt
@@ -4,8 +4,8 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Newspaper
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.mnhyim.home.destinations.CategoriesScreenDestination
-import com.mnhyim.home.destinations.SearchScreenDestination
+import com.mnhyim.news.destinations.CategoriesScreenDestination
+import com.mnhyim.news.destinations.SearchScreenDestination
 import com.ramcosta.composedestinations.spec.DirectionDestinationSpec
 
 enum class HomeNavigationBarDestinations(
