--- feature/home/src/main/java/com/mnhyim/home/categories/CategoriesItem.kt
@@ -1,47 +1,0 @@
-package com.mnhyim.home.categories
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Card
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CategoriesItem(
-    item: NewsCategories,
-    onClick: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        onClick = { onClick(item.category) },
-        modifier = modifier
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier.padding(32.dp)
-        ) {
-            Icon(
-                imageVector = item.icon,
-                contentDescription = "",
-                modifier = Modifier.fillMaxWidth()
-            )
-            Text(
-                text = "${item.category}",
-                textAlign = TextAlign.Center,
-                style = MaterialTheme.typography.labelSmall,
-                modifier = Modifier.fillMaxWidth().padding(top = 8.dp)
-            )
-        }
-    }
-}
