--- feature/home/src/main/java/com/mnhyim/home/categories/CategoriesScreen.kt
@@ -1,57 +1,0 @@
-package com.mnhyim.home.categories
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.mnhyim.home.destinations.SourcesScreenDestination
-import com.ramcosta.composedestinations.annotation.Destination
-import com.ramcosta.composedestinations.annotation.RootNavGraph
-import com.ramcosta.composedestinations.navigation.DestinationsNavigator
-
-@RootNavGraph(start = true)
-@Destination
-@Composable
-fun CategoriesScreen(
-    navigator: DestinationsNavigator
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier
-            .padding(8.dp)
-            .fillMaxSize()
-    ) {
-
-        Text(
-            text = "Categories",
-            textAlign = TextAlign.Center,
-            style = MaterialTheme.typography.headlineMedium
-        )
-
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(2),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier.padding(0.dp, 8.dp)
-        ) {
-            items(
-                items = NewsCategories.values(),
-            ) { categories ->
-                CategoriesItem(
-                    item = categories,
-                    onClick = { navigator.navigate(SourcesScreenDestination(category = it)) }
-                )
-            }
-        }
-    }
-}
