--- feature/home/src/main/java/com/mnhyim/home/components/OpenWebView.kt
@@ -1,50 +1,0 @@
-package com.mnhyim.home.components
-
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.viewinterop.AndroidView
-import com.ramcosta.composedestinations.annotation.Destination
-
-@Destination
-@Composable
-fun OpenWebView(url: String) {
-    val context = LocalContext.current
-    var isLoading by remember {
-        mutableStateOf(true)
-    }
-
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier.fillMaxSize()
-    ) {
-        AndroidView(
-            factory = {
-                WebView(context).apply {
-                    webViewClient = object : WebViewClient() {
-                        override fun onPageFinished(view: WebView?, url: String?) {
-                            super.onPageFinished(view, url)
-                            isLoading = false
-                        }
-                    }
-                    loadUrl(url)
-                    webViewClient
-                }
-            }
-        )
-        if (isLoading) {
-            CircularProgressIndicator()
-        }
-    }
-}
