--- feature/home/src/main/java/com/mnhyim/home/sources/SourcesViewModel.kt
@@ -1,20 +1,0 @@
-package com.mnhyim.home.sources
-
-import androidx.lifecycle.ViewModel
-import androidx.paging.PagingData
-import com.mnhyim.domain.model.sources.SourceModel
-import com.mnhyim.domain.usecase.NewsUseCases
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import javax.inject.Inject
-
-@HiltViewModel
-class SourcesViewModel @Inject constructor(
-    private val newsUseCases: NewsUseCases
-): ViewModel() {
-
-    fun getSources(category: String): Flow<PagingData<SourceModel>> = newsUseCases.getSourcesByCategoryUseCase(category)
-
-}
