+++ feature/news/src/main/AndroidManifest.xml
