+++ feature/news/src/main/java/com/mnhyim/news/news/NewsViewModel.kt
@@ -1,0 +1,19 @@
+package com.mnhyim.news.news
+
+import androidx.lifecycle.ViewModel
+import androidx.paging.PagingData
+import com.mnhyim.domain.model.articles.ArticleModel
+import com.mnhyim.domain.usecase.NewsUseCases
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+@HiltViewModel
+class NewsViewModel @Inject constructor(
+    private val newsUseCases: NewsUseCases
+): ViewModel() {
+
+
+    fun getNews(source: String): Flow<PagingData<ArticleModel>> = newsUseCases.getNewsBySourceUseCase(source)
+
+}
