+++ feature/news/src/main/java/com/mnhyim/news/sources/SourcesCardItem.kt
@@ -1,0 +1,65 @@
+package com.mnhyim.news.sources
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.mnhyim.domain.model.sources.SourceModel
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SourcesCardItem(
+    item: SourceModel,
+    onClick: (String, String, String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Card(
+        onClick = { onClick(item.id, item.name, item.description) },
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.primary,
+        ),
+        modifier = modifier,
+    ) {
+
+        Text(
+            text = item.name,
+            style = MaterialTheme.typography.labelLarge,
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(MaterialTheme.colorScheme.tertiary)
+                .padding(8.dp)
+        )
+
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(MaterialTheme.colorScheme.tertiary.copy(alpha = 0.75f))
+                .padding(8.dp, 4.dp)
+        ) {
+            Text(
+                text = item.country.uppercase(),
+                style = MaterialTheme.typography.bodySmall,
+            )
+            Spacer(modifier = Modifier.weight(1f))
+            Text(
+                text = item.language.uppercase(),
+                style = MaterialTheme.typography.bodySmall,
+            )
+        }
+
+        Text(
+            text = item.description,
+            style = MaterialTheme.typography.bodySmall,
+            modifier = Modifier.padding(12.dp)
+        )
+    }
+}
